+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 #1.0.2
+- Added option to show regular context menu if no action was selected
 - Now link/image/text selection menus can be populated with regular menu actions
 - It is now possible to deactive link/image/text selection menu by removing all of it's levels
 - Added few tooltips in options